+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Implement Modal Sheets for longer windows
     - Avoids UI elements getting under the dock
   - Add return to disk when selecting partitions
+  - Add "Search for disks again" option during OpenCore Install
 - Resolve failing to find binaries with `--patch_sys_vol` argument
 
 ## 0.4.5